+++ src/main/java/com/jrjackson/RubyAnySerializer.java
@@ -234,13 +234,13 @@ public class RubyAnySerializer extends JsonSerializer<IRubyObject> {
         if (df == null) {
             // DateFormat should always be set
             provider.defaultSerializeDateValue(dt.getJavaDate(), jgen);
-        } else if (df instanceof RubyDateFormat) {
+        } // RWB Note: I believe this is no longer used
+        else if (df instanceof RubyDateFormat) {
             // why another branch? I thought there was an easy win on to_s
             // maybe with jruby 9000
-            RubyDateFormat rdf = (RubyDateFormat) df.clone();
-            jgen.writeString(rdf.format(dt.getJavaDate()));
+            RubyDateFormat clonedRubyDateFormat = (RubyDateFormat) df.clone();
+            jgen.writeString(clonedRubyDateFormat.format(dt.getJavaDate()));
         } else {
-            SimpleDateFormat sdf = (SimpleDateFormat) df.clone();
             jgen.writeString(df.format(dt.getJavaDate()));
         }
     }